Replaces get_started with welcome function
